+++ backend/extensions.py
@@ -3,9 +3,11 @@
 from flask_sqlalchemy import SQLAlchemy
 from flask_bcrypt import Bcrypt
 from flask_jwt_extended import JWTManager
+from flask_migrate import Migrate
 
 # On crée les instances "vides" de nos extensions ici.
 # Elles seront configurées plus tard par notre create_app().
 db = SQLAlchemy()
 bcrypt = Bcrypt()
 jwt = JWTManager()
+migrate = Migrate()